Got connections canvas load to work
Nodeclass now handles the connections between node, and when trying to connect two nodes together, It will check to see if they already are connected, and if not it will add the connection for correct node as well as for the node its being connected too.

This is now much more consistent and far less likely to cause double connections going into the future.

I have not tested the search algorithems yet. I will do it next however