Significant progress
So I have nearly completed the intergration with depth first search.

Still tons of bugs to iron out like right now, it is claiming it found the goal despite clearly not.

However, I have found tons of things that needed tweaking that are now  working.

The biggest ones being the deletion of
 getNeighbors.

I added connectionPrinter, which I removed from addConnection to clean it up a little

which adds all the connections at the bottom of the screen. Plan on expanding this out to be alot more robust and so that it will color up what ever connections are turned on.

I made it so the nodes will now contain if they have been visited or not and if they are observed or not