fixed the air bubble step issue.
Fixed airbubble step issue. Now whenever we hit an already visited node, it skips it and goes to next step, however, I added "Was traversed" which means that the algo has gone over it.

This is because was_visited is more closer in meaning to "this is the current path"